LLE JIT: * Fixed a bug in the JCC instruction. * Changed the cycle count from u32 to u16. * Added cycle counting to the block linker. * Optimised the branch exit slightly.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@6681 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/DSPCore/Src/DSPCore.cpp
@@ -37,7 +37,7 @@
 SDSP g_dsp;
 DSPBreakpoints dsp_breakpoints;
 DSPCoreState core_state = DSPCORE_STOP;
-int cyclesLeft = 0;
+u16 cyclesLeft = 0;
 DSPEmitter *jit = NULL;
 Common::Event step_event;
